+++ fractal/components/stickies/StickyBar/StickyBar.config.js
@@ -4,6 +4,7 @@ module.exports = {
     "collated": true,
     "preview": "@stickies-preview",
     "context": {
-        "children": "Here is some text"
+        // "stickyPosition": "top",
+        "children": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam pretium vehicula imperdiet. Fusce vehicula volutpat nisl, mollis pellentesque arcu hendrerit in. Sed iaculis finibus lacus nec imperdiet. Duis feugiat, felis id venenatis consectetur, nisl tortor gravida ipsum, nec finibus nisl tellus in ex. Ut dignissim sem nisl, sed ultrices mi dignissim finibus."
     }
 };