+++ gulp/build-configs/webpack.config.style-guide.js
@@ -4,7 +4,8 @@ const webpack = require('webpack');
 module.exports = {
     target: 'web',
     entry: {
-        ToolTip: './fractal/components/tooltips/ToolTip/index.jsx'
+        ToolTip: './fractal/components/tooltips/ToolTip/index.jsx',
+        StickyBar: './fractal/components/stickies/StickyBar/index.jsx'
     },
     output: {
         path: path.resolve(__dirname, '../../_site/js'),
@@ -18,7 +19,7 @@ module.exports = {
                 use: {
                     loader: 'babel-loader',
                     options: {
-                        presets: ['react', 'env']
+                        presets: ['react', 'stage-0', 'env']
                     }
                 }
             }